Found one bug:
Ordering of filter result was set to Ascending. This would cause issues especially from the CLI when users specify a list of test files, they expect the test files to be executed in that order.

Hence, each pattern is a test file input parameter, and each bulk  of test files returned by each of those parameters will be executed,

This way, the user is in full control over the ordering of sets/subsets of test files